+++ .vscode/launch.json
@@ -21,7 +21,7 @@
 			"request": "launch",
 			"name": "Debug",
 			"program": "${workspaceFolder}/webserv",
-			"args": ["config/server.conf"],
+			"args": ["server.conf"],
 			"cwd": "${workspaceFolder}"
 		}
 	]
