+++ server.conf
@@ -5,7 +5,7 @@ http {
     # Server block 1
     server {
         listen 8080; # Port 80
-        server_name 127.0.0.1; # Hostname
+        server_name www.helloworld.com; # Hostname
 
         error_page 404 /404.html; # Default error pages
 
@@ -46,7 +46,7 @@ http {
 
     # Server block 2
     server {
-        listen 80; # Port 8080
+        listen 4242; # Port 8080
         server_name www.example.com; # Hostname Local host와 비교! http host header field와 비교 -> 일치하는 경우 처리, 없으면 default server block으로 처리
 
         # Redirect to a different URL
